requirements.txt
torch>=2.4.0
torchvision>=0.19.0
transformers>=4.46.0
accelerate>=0.26.0
opendatasets
pandas
scikit-learn
Pillow
scipy
grad-cam
matplotlib
tqdm
requests
opencv-python